linux-asahi: Remove custom rust wrappers
Upstream NixOS kernels are built with Rust support since some time now, so
there's no reason to continue rolling our own Rust support.

Instead of toggling the inclusion of the rust compiler in the build
environment based on the `withRust` parameter, we now conditionally
add the `CONFIG_RUST` config to the kernel, which triggers nixpkgs'
kernel expressions to include the required dependencies.